assignment completed

+++ src/controller/employee.controller.ts
@@ -2,29 +2,34 @@ import express from "express";
 import EmployeeService from "../service/employee.service";
 import HttpException from "../exceptions/http.exceptions";
 import { plainToInstance } from "class-transformer";
-import { CreateEmployeeDto } from "../dto/employee.dto";
+import { CreateEmployeeDto, UpdateEmployeeDto } from "../dto/employee.dto";
 import { validate } from "class-validator";
 import authorize from "../middleware/authorization.middleware";
 import Role from "../utils/role.enum";
+import { RequestWithUser } from "../utils/requestWithUser";
 
 class EmployeeController {
 	public router: express.Router;
 
 	constructor(private employeeService: EmployeeService) {
 		this.router = express.Router();
-		this.router.get("/", this.getAllEmployees);
-		this.router.get("/:id", this.getEmployeeById);
+		this.router.get("/", authorize, this.getAllEmployees);
+		this.router.get("/:id", authorize, this.getEmployeeById);
 		this.router.post("/", authorize, this.createEmployee);
-		this.router.delete("/:id", this.deleteEmployee);
-		this.router.put("/:id", this.updateEmployee);
+		this.router.delete("/:id", authorize, this.deleteEmployee);
+		this.router.put("/:id", authorize, this.updateEmployee);
 		this.router.post("/login", this.loginEmployee);
 	}
 	public getAllEmployees = async (
-		req: express.Request,
+		req: RequestWithUser,
 		res: express.Response,
 		next: express.NextFunction
 	) => {
 		try {
+			const role = req.role;
+			if (!role) {
+				throw new HttpException(403, "Please Login to Access");
+			}
 			const employees = await this.employeeService.getAllEmployees();
 			if (!employees) {
 				const error = new HttpException(404, `No Employees found`);
@@ -37,11 +42,15 @@ class EmployeeController {
 	};
 
 	public getEmployeeById = async (
-		req: express.Request,
+		req: RequestWithUser,
 		res: express.Response,
 		next: express.NextFunction
 	) => {
 		try {
+			const role = req.role;
+			if (!role) {
+				throw new HttpException(403, "Please Login to Access");
+			}
 			const employee = await this.employeeService.getEmployeeById(
 				Number(req.params.id)
 			);
@@ -59,12 +68,12 @@ class EmployeeController {
 	};
 
 	public createEmployee = async (
-		req: express.Request,
+		req: RequestWithUser,
 		res: express.Response,
 		next: express.NextFunction
 	) => {
 		try {
-			const role = req.body.role;
+			const role = req.role;
 			if (role != Role.HR) {
 				throw new HttpException(
 					403,
@@ -84,7 +93,8 @@ class EmployeeController {
 				employeeDto.age,
 				employeeDto.address,
 				employeeDto.password,
-				employeeDto.role
+				employeeDto.role,
+				employeeDto.department
 			);
 			res.status(201).send(employee);
 		} catch (error) {
@@ -93,11 +103,19 @@ class EmployeeController {
 	};
 
 	public deleteEmployee = async (
-		req: express.Request,
+		req: RequestWithUser,
 		res: express.Response,
 		next: express.NextFunction
 	) => {
 		try {
+			const role = req.role;
+			if (role != Role.HR) {
+				throw new HttpException(
+					403,
+					"You are not authorized to delete employee"
+				);
+			}
+
 			const result = await this.employeeService.deleteEmployee(
 				Number(req.params.id)
 			);
@@ -110,15 +128,32 @@ class EmployeeController {
 	};
 
 	public updateEmployee = async (
-		req: express.Request,
+		req: RequestWithUser,
 		res: express.Response,
 		next: express.NextFunction
 	) => {
-		const employee = await this.employeeService.updateEmployee(
-			Number(req.params.id),
-			req.body
-		);
-		res.status(200).send(employee);
+		try {
+			const role = req.role;
+			if (role != Role.HR) {
+				throw new HttpException(
+					403,
+					"You are not authorized to update employee"
+				);
+			}
+			const employeeDto = plainToInstance(UpdateEmployeeDto, req.body);
+			const errors = await validate(employeeDto);
+			if (errors.length) {
+				console.log(errors);
+				throw new HttpException(400, JSON.stringify(errors));
+			}
+			const employee = await this.employeeService.updateEmployee(
+				Number(req.params.id),
+				req.body
+			);
+			res.status(200).send(employee);
+		} catch (error) {
+			next(error);
+		}
 	};
 
 	public loginEmployee = async (
@@ -132,6 +167,9 @@ class EmployeeController {
 				email,
 				password
 			);
+			if (!token) {
+				throw new HttpException(404, "Inavalid Credentials.");
+			}
 			res.status(200).send(token);
 		} catch (error) {
 			next(error);
